+++ crates/holochain_keystore/src/crude_mock_keystore.rs
@@ -2,14 +2,12 @@
 //! call. This is about as close as we can get to a true mock which would allow
 //! tweaking individual handlers, hence why this is a "crude" mock.
 
-use crate::spawn_test_keystore;
 use crate::MetaLairClient;
 use futures::FutureExt;
 use lair_keystore::dependencies::lair_keystore_api::lair_client::client_traits::AsLairClient;
 use lair_keystore::dependencies::lair_keystore_api::prelude::{LairApiEnum, LairClient};
 use lair_keystore::dependencies::lair_keystore_api::types::SharedSizedLockedArray;
 use lair_keystore::dependencies::lair_keystore_api::LairResult;
-use std::sync::atomic::AtomicBool;
 use std::sync::Arc;
 
 /// Spawn a test keystore which always returns the same LairError for every call.
@@ -26,61 +24,6 @@ where
     )
 }
 
-/// Spawn a test keystore that can switch between mocked and real.
-/// It starts off as real and can be switched to the given callback mock
-/// using the [`MockLairControl`].
-pub async fn spawn_real_or_mock_keystore<F>(
-    func: F,
-) -> LairResult<(MetaLairClient, MockLairControl)>
-where
-    F: Fn(LairApiEnum) -> LairResult<LairApiEnum> + Send + Sync + 'static,
-{
-    let real = spawn_test_keystore().await?;
-    let use_mock = Arc::new(AtomicBool::new(false));
-    let mock = RealOrMockKeystore {
-        mock: Box::new(func),
-        real,
-        use_mock: use_mock.clone(),
-    };
-
-    let control = MockLairControl(use_mock);
-
-    let (s, _) = tokio::sync::mpsc::unbounded_channel();
-    Ok((
-        MetaLairClient(
-            Arc::new(parking_lot::Mutex::new(LairClient(Arc::new(mock)))),
-            s,
-        ),
-        control,
-    ))
-}
-/// A keystore which always returns the same LairError for every call.
-struct RealOrMockKeystore {
-    mock: Box<dyn Fn(LairApiEnum) -> LairResult<LairApiEnum> + Send + Sync + 'static>,
-    real: MetaLairClient,
-    use_mock: Arc<AtomicBool>,
-}
-
-/// Control if a mocked lair keystore is using
-/// the real keystore or the mock callback.
-pub struct MockLairControl(Arc<AtomicBool>);
-
-impl MockLairControl {
-    /// Use the mock callback.
-    pub fn use_mock(&self) {
-        self.0.store(true, std::sync::atomic::Ordering::SeqCst);
-    }
-
-    /// Use the real test keystore.
-    pub fn use_real(&self) {
-        self.0.store(false, std::sync::atomic::Ordering::SeqCst);
-    }
-
-    /// Is the keystore using the mock?
-    pub fn using_mock(&self) -> bool {
-        self.0.load(std::sync::atomic::Ordering::SeqCst)
-    }
-}
 /// A keystore which always returns the same LairError for every call.
 struct CrudeMockKeystore(Arc<dyn Fn() -> one_err::OneErr + Send + Sync + 'static>);
 
@@ -103,32 +46,6 @@ impl AsLairClient for CrudeMockKeystore {
     ) -> futures::future::BoxFuture<'static, LairResult<LairApiEnum>> {
         let err = (self.0)();
         async move { Err(err) }.boxed()
-    }
-}
-
-impl AsLairClient for RealOrMockKeystore {
-    fn get_enc_ctx_key(&self) -> SharedSizedLockedArray<32> {
-        self.real.cli().0.get_enc_ctx_key()
-    }
-
-    fn get_dec_ctx_key(&self) -> SharedSizedLockedArray<32> {
-        self.real.cli().0.get_dec_ctx_key()
-    }
-
-    fn shutdown(&self) -> futures::future::BoxFuture<'static, LairResult<()>> {
-        self.real.cli().0.shutdown().boxed()
-    }
-
-    fn request(
-        &self,
-        request: LairApiEnum,
-    ) -> futures::future::BoxFuture<'static, LairResult<LairApiEnum>> {
-        if self.use_mock.load(std::sync::atomic::Ordering::SeqCst) {
-            let r = (self.mock)(request);
-            async move { r }.boxed()
-        } else {
-            AsLairClient::request(&*self.real.cli().0 .0, request)
-        }
     }
 }