+++ crates/holochain_zome_types/src/zome_io.rs
@@ -128,9 +128,6 @@ wasm_io_types! {
     #[cfg(feature = "unstable-functions")]
     fn sleep (core::time::Duration) -> ();
 
-    // @todo
-    fn version (()) -> zt::version::ZomeApiVersion;
-
     // Attempt to have the keystore sign some data
     // The pubkey in the input needs to be found in the keystore for this to work
     fn sign (zt::signature::Sign) -> zt::signature::Signature;